fix: cleanup voting on pm proposal page
